+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Support streaming uploads for services like S3
 - Implement `DispatchSignedRequest` and `ProvideAwsCredentials` for `Arc<>` and `Rc<>` wrapped types
 - Add CodeDeploy Commands service.
+- Changed HttpClient to be generic over hyper::Connect, default HttpClient<C = HttpsConnector<HttpConnector>>, see #1033
+- Add CodeDeploy Commands service.
 
 ## [0.32.0] - 2018-03-03